Create activity_add_family.xml
The XML file activity_add_family.xml, which contains the arrangement for adding family members, is included with this commit. In addition to a Save button, the layout has EditText areas where users can enter up to three cellphone numbers. In order to improve the functionality of the application, family members' contact information will be captured using this layout.