Created LoginActivity.java
This commit adds the Java class "LoginActivity.java", which provides functionality for user login. The class includes methods to handle user input validation, authentication, and session management. Upon successful login, the user is directed to either the AdminHome  (if the username and password are "admin"). This class enhances the application by enabling users to securely log in and access relevant features based on their role.