Create Activity_view_user.xml
A new XML layout file for showing user information in an Android app is introduced in this patch. The layout has buttons to delete the user and cancel the action, as well as Text Views for the name, email, cellphone number, address, and description. A Text View is also offered for further information. The arrangement of the software makes it easier for users to engage and administer it."