Create ViewUser.java
The ViewUser activity class, which enables users to access comprehensive details about a particular user, is included in this commit. The activity uses the user ID supplied via the intent to retrieve user information from the Firebase Realtime Database. Name, email, phone number, address, and description of the user are displayed. The delete button might not be available depending on the user's role (admin or not). If allowed, users can cancel the activity to go back to the relevant home screen or delete the user. This approach improves the application's user management."